+++ Config/BuildSettings.swift
@@ -125,7 +125,7 @@ final class BuildSettings: NSObject {
     
     // MARK: - VoIP
     static var allowVoIPUsage: Bool {
-        #if canImport(JitsiMeet)
+        #if canImport(JitsiMeetSDK)
         return true
         #else
         return false